% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_collation.R
\name{pg_collation}
\alias{pg_collation}
\title{Create non-deterministic collation}
\usage{
pg_collation(
  conn,
  collation_name = NULL,
  locale = "und",
  provider = "icu",
  numeric = FALSE,
  case_insensitive = FALSE,
  ignore_accents = FALSE,
  alternate_shifted = FALSE,
  uppercase_first = FALSE
)
}
\arguments{
\item{conn}{Connection}

\item{collation_name}{Collation name}

\item{locale}{Locale. Defaults to none = "und-u"}

\item{provider}{Privider. Defaults to "icu". On Unix systems, you can
get the available locales with the shell command locale -a}

\item{numeric}{Numeric ordering, sorts sequences of digits by their numeric value}

\item{case_insensitive}{Ignore case?}

\item{ignore_accents}{Ignore accents}

\item{alternate_shifted}{Ignore punctuation?}

\item{uppercase_first}{Sorts uppercase letter before lowercase letters.}
}
\value{
Dataframe
}
\description{
Create non-deterministic collation
}
\examples{
\dontrun{
pg_collation(conn,
            collation_name = "coll_phone",
            locale = 'en',
            provider = 'icu',
            numeric = TRUE,
            )
}
}
