% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_d.R
\name{pg_d}
\alias{pg_d}
\title{List tables or table columns}
\usage{
pg_d(conn, plus = FALSE, tbl = NULL, schema = "public")
}
\arguments{
\item{conn}{Connection}

\item{plus}{Additional information?}

\item{tbl}{Table. If informed, describes table columns}

\item{schema}{Default to public}
}
\value{
Table
}
\description{
List tables or table columns
}
\examples{
\dontrun{
pg_d(conn)
}
}
