% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_da.R
\name{pg_dA}
\alias{pg_dA}
\title{List access methods}
\usage{
pg_dA(conn)
}
\arguments{
\item{conn}{Connection}
}
\value{
Dataframe with access methods
}
\description{
List access methods
}
\examples{
\dontrun{
pg_dA(conn)
}
}
