% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_da.R
\name{pg_dAc}
\alias{pg_dAc}
\title{List Operarator classes}
\usage{
pg_dAc(conn)
}
\arguments{
\item{conn}{Connection}
}
\value{
Dataframe with operators' classes
}
\description{
List Operarator classes
}
\examples{
\dontrun{
pg_dAc(conn)
}

}
