% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_da.R
\name{pg_dAcp}
\alias{pg_dAcp}
\title{List operators' classes with additional information}
\usage{
pg_dAcp(conn)
}
\arguments{
\item{conn}{Connection}
}
\value{
Dataframe
}
\description{
List operators' classes with additional information
}
\examples{
\dontrun{
pg_dAcp(conn)
}
}
