% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_da.R
\name{pg_dAp}
\alias{pg_dAp}
\title{List access methods with additional information}
\usage{
pg_dAp(conn)
}
\arguments{
\item{conn}{Connection}
}
\value{
Dataframe with access methods
}
\description{
List access methods with additional information
}
\examples{
\dontrun{
pg_dAp(conn)
}
}
