% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_size.R
\name{pg_db_size}
\alias{pg_db_size}
\title{Gets database size}
\usage{
pg_db_size(conn, dbname = NULL)
}
\arguments{
\item{conn}{Connection}

\item{dbname}{Database name. If not provided, it returns the current
database size.}
}
\value{
Named string with database size
}
\description{
Gets database size
}
\examples{
\dontrun{
pg_db_size(conn,"postgres")
}
}
