% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_df.R
\name{pg_dF}
\alias{pg_dF}
\title{List text search configuration}
\usage{
pg_dF(conn, name = NULL)
}
\arguments{
\item{conn}{Connection}

\item{name}{Optional. Configuration name}
}
\value{
Tibble with columns schema, name, and descriptio
}
\description{
List text search configuration
}
\examples{
\dontrun{
dF(conn)
}
}
