% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_dn.R
\name{pg_dn}
\alias{pg_dn}
\title{Describe schemas}
\usage{
pg_dn(conn, schema = NULL)
}
\arguments{
\item{conn}{Connection}

\item{schema}{Optional. If you want to describe just
one schema}
}
\value{
data.frame with shema names and owners
}
\description{
Describe schemas
}
\examples{
\dontrun{
pg_dn(conn)
}
}
