% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_dn.R
\name{pg_dnp}
\alias{pg_dnp}
\title{Describe schemas plus}
\usage{
pg_dnp(conn, schema = NULL)
}
\arguments{
\item{conn}{Connections}

\item{schema}{Optional. Schema name}
}
\value{
Data.frame with 4 columns: name, owner, access_privileges,
    and description
}
\description{
Describe schemas plus
}
\examples{
\dontrun{
pg_dnp(conn)
}
}
