% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_du.R
\name{pg_du}
\alias{pg_du}
\title{Lists roles and their privileges}
\usage{
pg_du(conn, plus = FALSE, user = NULL)
}
\arguments{
\item{conn}{Connection}

\item{plus}{Add description?}

\item{user}{Optional: username}
}
\value{
tibble
}
\description{
Lists roles and their privileges
}
