% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_fts.R
\name{pg_fts}
\alias{pg_fts}
\title{Creates text search configuration}
\usage{
pg_fts(
  conn,
  name,
  source_config = "english",
  mapping = list(parsers = NULL, with = NULL)
)
}
\arguments{
\item{conn}{Connection}

\item{name}{The name of the text search configuration to be created.
The name can be schema-qualified.}

\item{source_config}{The name of an existing text search configuration to copy.
It defaults to English.}

\item{mapping}{A list of parsers}
}
\value{
Returns 0 if accomplished.
}
\description{
Creates text search configuration
}
