% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_l.R
\name{pg_l}
\alias{pg_l}
\title{List all databases or a specific one}
\usage{
pg_l(conn, dbname = NULL)
}
\arguments{
\item{conn}{Connection}

\item{dbname}{Optional database name}
}
\value{
data.frame
}
\description{
List all databases or a specific one
}
\examples{
\dontrun{
pg_l(conn)
}
}
