% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_list_connections.R
\name{pg_list_connections}
\alias{pg_list_connections}
\title{List connections}
\usage{
pg_list_connections(conn)
}
\arguments{
\item{conn}{Your connection}
}
\value{
Data.frame with information about current connections
}
\description{
List connections
}
\examples{
\dontrun{
pg_list_connections(conn= conn)
}
}
