% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_l.R
\name{pg_lp}
\alias{pg_lp}
\title{List databases with addicional information}
\usage{
pg_lp(conn, dbname = NULL)
}
\arguments{
\item{conn}{Connection}

\item{dbname}{Optional database's name}
}
\value{
Dataframe
}
\description{
List databases with addicional information
}
\examples{
\dontrun{
pg_lp(conn)
}
}
