% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_pivot_longer.R
\name{pg_pivot_longer}
\alias{pg_pivot_longer}
\title{Convert PostgreSQL table to long format}
\usage{
pg_pivot_longer(conn, tbl, new_tbl = NULL, cols, names_to, values_to)
}
\arguments{
\item{conn}{Connection.}

\item{tbl}{Database table.}

\item{new_tbl}{Optional. If you want to create a new table on the database
intead of importing to R.}

\item{cols}{<tidy-select> Variables to pivot into long format.}

\item{names_to}{A string specifying the name of the column to create from
the data stored in the column names of tbl.}

\item{values_to}{A string specifying the name of the column to create
from the data stored in cell values.}
}
\value{
tibbl or a table in the database
}
\description{
Convert PostgreSQL table to long format
}
