% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_summary.R
\name{pg_summary}
\alias{pg_summary}
\title{Creates a summary table from PostgreSQL table's columns}
\usage{
pg_summary(
  conn,
  tbl,
  ...,
  schema = "public",
  max_levels = 10,
  outliers = FALSE
)
}
\arguments{
\item{conn}{Connection}

\item{tbl}{Table name}

\item{...}{Columns' bare-names}

\item{schema}{Schema name, "public" by default}

\item{max_levels}{Max number of levels to report for categorical}

\item{outliers}{Calculate outlier for numeric data? Defaults to FALSE.}
}
\value{
Dataframe
}
\description{
Creates a summary table from PostgreSQL table's columns
}
\details{
R exports factors as text to PostgreSQL, even though PostgreSQL has enum type
    which is somehow equivalent to factors.  This function shows unique values as
    variable levels to maximum number of 10, unless otherwise specified.
    As of now, it f supports only dates, texts and numbers.
}
