% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_size.R
\name{pg_tbl_size}
\alias{pg_tbl_size}
\title{Gets table's size}
\usage{
pg_tbl_size(conn, tbl = NULL)
}
\arguments{
\item{conn}{Connection}

\item{tbl}{String with table's name}
}
\value{
Named string with table's size
}
\description{
Gets table's size
}
\examples{
\dontrun{
pg_tbl_size(conn,"data")
}
}
