% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_terminate_backend.R
\name{pg_terminate_backend}
\alias{pg_terminate_backend}
\title{Terminate one or all connections by pid}
\usage{
pg_terminate_backend(conn, pid = NULL)
}
\arguments{
\item{conn}{Connection}

\item{pid}{NULL, defaults to all connections to databases,
 except the your own, or
the connection pid you want to terminate.}
}
\value{
data.frame with one row per terminated connection
}
\description{
Terminate one or all connections by pid
}
\details{
You can list all connections with ```pg_list_connections()```
}
\examples{
\dontrun{
pg_terminate_backend(conn)
}
}
