% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_d.R
\name{pg_d}
\alias{pg_d}
\title{List tables or table columns}
\usage{
pg_d(conn, tbl = NULL, plus = FALSE)
}
\arguments{
\item{conn}{Connection}

\item{tbl}{Character. If informed, describes table columns}

\item{plus}{Additional information?}
}
\value{
Tibble
}
\description{
List tables or table columns
}
\examples{
\dontrun{
pg_d(conn)
}
}
